art_sensei/requirements.txt
fastapi
uvicorn[standard]
google-generativeai
python-dotenv
pydantic-settings
requests
deepgram-sdk>=3.0
certifi
